Add and use display name variable
Parabolic is currently listed in all-lowercase as «parabolic» in app
stores, which is not a particularly flattering look.

This PR adds a separate variable containing a «pretty» display version
of the name. I'm not well-versed in CMake at all, but I suppose there
isn't more that needs to be done for it to work.